DEV: Remove buffered rendering from input tips
This is another refactoring in the multi-step process to remove all uses
of our custom Render Buffer.

Previous commit: 2290ec9e87ab005fcb6666c1448bdbe4e7b92bd6 in this
series.

This commit affects the display of input tips. It is just a refactor and
does not change any functionality.